move all tests under Adapter namespace, add more tests
