+++ tests/phpunit/Adapter/CodeceptionAdapterFactoryTest.php
@@ -33,7 +33,7 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\TestFramework\Codeception;
+namespace Infection\Tests\TestFramework\Codeception\Adapter;
 
 use Infection\TestFramework\Codeception\CodeceptionAdapter;
 use Infection\TestFramework\Codeception\CodeceptionAdapterFactory;
@@ -50,7 +50,7 @@ final class CodeceptionAdapterFactoryTest extends TestCase
         $adapter = CodeceptionAdapterFactory::create(
             '/path/to/codecept',
             '/tmp',
-            __DIR__ . '/Fixtures/Files/codeception/codeception.yml',
+            __DIR__ . '/../Fixtures/Files/codeception/codeception.yml',
             null,
             '/path/to/junit.xml',
             '/path/to/project',
@@ -78,7 +78,7 @@ final class CodeceptionAdapterFactoryTest extends TestCase
         CodeceptionAdapterFactory::create(
             '/path/to/codecept',
             '/tmp',
-            __DIR__ . '/Fixtures/Files/codeception/invalid_codeception.yml',
+            __DIR__ . '/../Fixtures/Files/codeception/invalid_codeception.yml',
             null,
             '/path/to/junit.xml',
             '/path/to/project',
