+++ tests/phpunit/Adapter/VersionParserTest.php
@@ -33,7 +33,7 @@
 
 declare(strict_types=1);
 
-namespace Infection\Tests\TestFramework\Codeception;
+namespace Infection\Tests\TestFramework\Codeception\Adapter;
 
 use Infection\TestFramework\Codeception\VersionParser;
 use InvalidArgumentException;
